refactor: standardize function names, enhance extraction logic, and improve test coverage
