+++ internal/parsers/awards.go
@@ -4,11 +4,22 @@ import (
 	"regexp"
 	"strings"
 
+	"github.com/gocolly/colly/v2"
 	"github.com/ngshiheng/michelin-my-maps/v3/internal/models"
 )
 
+// ExtractDistinction extracts the Michelin distinction and green star status from a restaurant HTML element.
+// Returns (distinction string, greenStar bool).
+func ExtractDistinction(e *colly.XMLElement) (string, bool) {
+	distinction := tryAwardSelectors(e, "distinction", parseDistinction)
+	if distinction == "" {
+		distinction = models.SelectedRestaurants
+	}
+	greenStar := tryAwardSelectors(e, "greenStar", parseGreenStar) == "true"
+	return distinction, greenStar
+}
+
 var (
-	// Michelin distinction patterns for parsing award levels
 	re3Stars      = regexp.MustCompile(`(?i)\b(three|3)\b.*?\bstars?\b`)
 	re2Stars      = regexp.MustCompile(`(?i)\b(two|2)\b.*?\bstars?\b`)
 	re1Star       = regexp.MustCompile(`(?i)\b(one|1)\b.*?\bstar\b`)
@@ -16,16 +27,16 @@ var (
 	reSelected    = regexp.MustCompile(`(?i)\bselected\s*restaurants?\b|\bplate\b`)
 )
 
-func ParseGreenStar(text string) string {
+func parseGreenStar(text string) string {
 	if strings.Contains(strings.ToLower(text), "green star") {
 		return "true"
 	}
 	return "false"
 }
 
-func ParseDistinction(text string) string {
+func parseDistinction(text string) string {
 	distinction := strings.ToLower(text)
-	distinction = decodeHTMLEntities(distinction)
+	distinction = DecodeHTMLEntities(distinction)
 	distinction = strings.Trim(distinction, " .!?,;:-")
 	distinction = strings.TrimSpace(distinction)
 
@@ -43,10 +54,4 @@ func ParseDistinction(text string) string {
 	default:
 		return models.SelectedRestaurants
 	}
-}
-
-func decodeHTMLEntities(text string) string {
-	text = strings.ReplaceAll(text, "&bull;", "")
-	text = strings.ReplaceAll(text, "•", "")
-	return text
 }
