+++ internal/parsers/phonenumber_test.go
@@ -14,7 +14,7 @@ func TestParsePhoneNumber(t *testing.T) {
 	}
 	for _, tt := range tests {
 		t.Run(tt.input, func(t *testing.T) {
-			got := ParsePhoneNumber(tt.input)
+			got := ExtractPhoneNumber(tt.input)
 			if got != tt.expected {
 				t.Errorf("ParsePhoneNumber(%q) = %q; want %q", tt.input, got, tt.expected)
 			}
